requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.33.0
pydantic-settings==2.7.1
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
openai==1.59.7
APScheduler==3.11.0
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.20
